+++ src/components/connections/ConnectionTable.vue
@@ -192,8 +192,7 @@
 </template>
 
 <script setup lang="ts">
-import { disconnectByIdAPI } from '@/api'
-import { blockconnectByIdAPI } from '@/api'
+import { blockconnectByIdAPI, disconnectByIdAPI } from '@/api'
 import { useConnections } from '@/composables/connections'
 import {
   CONNECTION_TAB_TYPE,
@@ -229,6 +228,7 @@ import {
   MagnifyingGlassMinusIcon,
   MagnifyingGlassPlusIcon,
   MapPinIcon,
+  NoSymbolIcon,
   XMarkIcon,
 } from '@heroicons/vue/24/outline'
 import {
@@ -302,25 +302,29 @@ const columns: ColumnDef<Connection>[] = [
         ],
       )
 
-      const degradeButton = row.original.metadata.smartBlock === 'normal' ? h(
-        'button',
-        {
-          class: 'btn btn-xs btn-circle',
-          onClick: (e) => {
-            const connection = row.original
+      if (row.original.metadata.smartBlock === 'normal') {
+        const degradeButton = h(
+          'button',
+          {
+            class: 'btn btn-xs btn-circle',
+            onClick: (e) => {
+              const connection = row.original
 
-            e.stopPropagation()
-            blockconnectByIdAPI(connection.id)
+              e.stopPropagation()
+              blockconnectByIdAPI(connection.id)
+            },
           },
-        },
-        [
-          h(NoSymbolIcon, {
-            class: 'h-4 w-4',
-          }),
-        ],
-      ) : null
+          [
+            h(NoSymbolIcon, {
+              class: 'h-4 w-4',
+            }),
+          ],
+        )
 
-      return h('div', { class: 'flex gap-1' }, [closeButton, degradeButton].filter(Boolean))
+        return h('div', { class: 'flex gap-1' }, [closeButton, degradeButton])
+      }
+
+      return closeButton
     },
   },
   {