+++ src/components/sidebar/ConnectionCtrl.tsx
@@ -1,5 +1,5 @@
 import { disconnectAllAPI, disconnectByIdAPI } from '@/api'
-import { SORT_DIRECTION, SORT_TYPE } from '@/constant'
+import { ROUTE_NAME, SORT_DIRECTION, SORT_TYPE } from '@/constant'
 import { useTooltip } from '@/helper/tooltip'
 import {
   connectionFilter,
@@ -23,6 +23,7 @@ import {
 } from '@heroicons/vue/24/outline'
 import { defineComponent, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
+import { useRouter } from 'vue-router'
 import DialogWrapper from '../common/DialogWrapper.vue'
 import TextInput from '../common/TextInput.vue'
 import ConnectionCardSettings from '../settings/ConnectionCardSettings.vue'
@@ -55,16 +56,19 @@ export default defineComponent({
   },
   setup(props) {
     const { t } = useI18n()
+    const router = useRouter()
     const settingsModel = ref(false)
     const { showTip } = useTooltip()
 
     return () => {
       const sortForCards = (
-        <div class="flex w-full items-center gap-1 text-sm lg:w-auto">
+        <div
+          class={`flex items-center gap-1 text-sm ${props.isLargeCtrlsBar ? 'w-auto' : 'w-full'}`}
+        >
           <span class="shrink-0">{t('sortBy')}</span>
-          <div class="join flex-1 max-lg:w-0">
+          <div class={`join flex-1 ${props.isLargeCtrlsBar ? 'min-w-46' : ''}`}>
             <select
-              class="join-item select select-sm flex-1 max-lg:w-0"
+              class="join-item select select-sm flex-1"
               v-model={connectionSortType.value}
             >
               {(Object.values(SORT_TYPE) as string[]).map((opt) => (
@@ -103,7 +107,10 @@ export default defineComponent({
           >
             <WrenchScrewdriverIcon class="h-4 w-4" />
           </button>
-          <DialogWrapper v-model={settingsModel.value}>
+          <DialogWrapper
+            v-model={settingsModel.value}
+            title={t('connectionSettings')}
+          >
             <div class="flex flex-col gap-4 p-2 text-sm">
               <div class="flex items-center gap-2">
                 <span class="shrink-0">{t('hideConnectionRegex')}</span>
@@ -130,6 +137,19 @@ export default defineComponent({
                 </div>
               </div>
               {useConnectionCard.value ? <ConnectionCardSettings /> : <TableSettings />}
+              <div class="divider m-0"></div>
+              <button
+                class="btn btn-block"
+                onClick={() => {
+                  settingsModel.value = false
+                  router.push({
+                    name: ROUTE_NAME.settings,
+                    query: { scrollTo: 'connections-settings' },
+                  })
+                }}
+              >
+                {t('moreSettings')}
+              </button>
             </div>
           </DialogWrapper>
         </>